requirements.txt
Flask==2.2.5
pandas==2.1.4
python-docx==1.1.2
